+++ triple_arena/CHANGELOG.md
@@ -4,9 +4,11 @@
 ### Additions
 - Added `LinkNoGen` and `ChainNoGenArena` and made various performance improvements
 - Added `SurjectArena::get_link_no_gen`
+- Added `OrdArena::get_link_no_gen`
 
 ### Changes
 - Renamed all `*_ignore_gen` functions to `*_no_gen`
+- Changed the signatures of various `OrdArena` functions
 
 ## [0.11.0] - 2023-08-20
 ### Fixes